move back api trait at root of the crate
