+++ primitives/rpc/debug/src/lib.rs
@@ -24,6 +24,47 @@ pub mod api;
 pub mod v1;
 pub mod v2;
 
+use crate::api::*;
+use ethereum::TransactionV0 as Transaction;
+use sp_std::vec::Vec;
+
+sp_api::decl_runtime_apis! {
+	#[api_version(3)]
+	pub trait DebugRuntimeApi {
+
+		#[changed_in(2)]
+		fn trace_transaction(
+			extrinsics: Vec<Block::Extrinsic>,
+			transaction: &Transaction,
+			trace_type: single::TraceType,
+		) -> Result<single::TransactionTrace, sp_runtime::DispatchError>;
+
+		#[changed_in(2)]
+		fn trace_block(
+			extrinsics: Vec<Block::Extrinsic>,
+		) -> Result<Vec<block::TransactionTrace>, sp_runtime::DispatchError>;
+
+		#[changed_in(3)]
+		fn trace_transaction(
+			header: &Block::Header,
+			extrinsics: Vec<Block::Extrinsic>,
+			transaction: &Transaction,
+			trace_type: single::TraceType,
+		) -> Result<(), sp_runtime::DispatchError>;
+
+		fn trace_transaction(
+			header: &Block::Header,
+			extrinsics: Vec<Block::Extrinsic>,
+			transaction: &Transaction,
+		) -> Result<(), sp_runtime::DispatchError>;
+
+		fn trace_block(
+			header: &Block::Header,
+			extrinsics: Vec<Block::Extrinsic>,
+		) -> Result<(), sp_runtime::DispatchError>;
+	}
+}
+
 /// Runtime api closure result.
 #[derive(Debug)]
 pub enum Response {
